3.-RESTful-API/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-filter>=23.0.0
psycopg2-binary>=2.9.0
